rename existing Sequence to SequenceQuality
